move generate upload in pending folders to background
